Implement --max switch to limit commits processed per run
For very large code bases, --limit and --commits-per-day are good starting
points for gradually analysing the codebase, but they require constant
tweaking while the analysis gradually happens.

Instead this implements a --max X switch that will analyse up to X
unanalysed commits (of the last --limit commits, if necessary).

This means that if the build can analyse e.g. 30 commits before the job
times out, then we can always maximise this 30 commit window for each
analysis, and we can always gradually improve the resolution of our
analysis.

This also moves the db/schema.rb to a migration, so that we can
run multiple integration tests successfully.