Switch order of `git checkout master` and `git pull`
For repositories that had already been partially analysed, it seems
the order of `git pull origin master` and then `git checkout master`
was not actually updating the local branch to the latest master.

Switching these commands around should continue to work and should
also correctly update the local branch to the latest master.